browser-use/requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.76.0
anyio==4.12.1
attrs==26.1.0
backoff==2.2.1
beautifulsoup4==4.14.3
browser-use==0.12.6
browser-use-sdk==2.0.15
bubus==1.5.6
cdp-use==1.4.5
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.1
cloudpickle==3.1.2
cryptography==46.0.7
distro==1.9.0
docstring-parser==0.18.0
frozenlist==1.8.0
google-api-core==2.29.0
google-api-python-client==2.188.0
google-auth==2.48.0
google-auth-httplib2==0.3.1
google-auth-oauthlib==1.2.4
google-genai==1.65.0
googleapis-common-protos==1.74.0
groq==1.0.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
httpx-sse==0.4.3
idna==3.12
inquirerpy==0.3.4
jiter==0.14.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
lxml==6.1.0
markdown-it-py==4.0.0
markdownify==1.2.2
mcp==1.26.0
mdurl==0.1.2
multidict==6.7.1
oauthlib==3.3.1
ollama==0.6.1
openai==2.16.0
pfzy==0.3.4
pillow==12.1.1
portalocker==3.2.0
posthog==7.7.0
prompt-toolkit==3.0.52
propcache==0.4.1
proto-plus==1.27.2
protobuf==6.33.6
psutil==7.2.2
pyasn1==0.6.3
pyasn1-modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-settings==2.14.0
pygments==2.20.0
pyjwt==2.12.1
pyotp==2.9.0
pyparsing==3.3.2
pypdf==6.9.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-multipart==0.0.26
referencing==0.37.0
reportlab==4.4.9
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.3.1
rpds-py==0.30.0
rsa==4.9.1
screeninfo==0.8.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
sse-starlette==3.3.4
starlette==1.0.0
tenacity==9.1.4
tqdm==4.67.3
typing-extensions==4.15.0
typing-inspection==0.4.2
uritemplate==4.2.0
urllib3==2.6.3
uuid7==0.1.0
uvicorn==0.45.0
wcwidth==0.6.0
websockets==16.0
yarl==1.23.0
